rename : remove parent package
