--- src/main/java/com/kickball/domain/member/domain/RoleType.java
@@ -1,5 +1,0 @@
-package com.kickball.domain.member.domain;
-
-public enum RoleType {
-	USER, ADMIN,UNKNOWN;
-}
