--- src/main/java/com/kickball/domain/reservation/repository/ReservationRepository.java
@@ -1,8 +1,0 @@
-package com.kickball.domain.reservation.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.kickball.domain.reservation.domain.Reservation;
-
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-}
