--- src/main/java/com/kickball/domain/stadium/repository/StadiumRepository.java
@@ -1,8 +1,0 @@
-package com.kickball.domain.stadium.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.kickball.domain.stadium.domain.Stadium;
-
-public interface StadiumRepository extends JpaRepository<Stadium, Long> {
-}
