+++ src/main/java/com/kickball/member/domain/Member.java
@@ -1,0 +1,44 @@
+package com.kickball.member.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@Table(name = "members")
+@Entity
+public class Member {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column(name = "username", nullable = false, length = 50)
+	private String username;
+
+	@Column(name = "password", nullable = false, length = 100)
+	private String password;
+
+	@Enumerated(EnumType.STRING)
+	@Column(name = "role", nullable = false, length = 50)
+	private RoleType role;
+
+	@Column(name = "provider", nullable = false, length = 50)
+	private String provider;
+
+	public Member(String username, String password, String provider) {
+		this.username = username;
+		this.password = password;
+		this.role = RoleType.USER;
+		this.provider = provider;
+	}
+}
