+++ src/main/java/com/kickball/member/domain/RoleType.java
@@ -1,0 +1,5 @@
+package com.kickball.member.domain;
+
+public enum RoleType {
+	USER, ADMIN,UNKNOWN;
+}
