+++ src/main/java/com/kickball/reservation/repository/ReservationRepository.java
@@ -1,0 +1,8 @@
+package com.kickball.reservation.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.kickball.reservation.domain.Reservation;
+
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+}
