+++ src/main/java/com/kickball/stadium/dto/response/StadiumResponse.java
@@ -1,0 +1,20 @@
+package com.kickball.stadium.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class StadiumResponse {
+
+	private Long id;
+
+	public static StadiumResponse toDto(Long id) {
+		return new StadiumResponse(id);
+	}
+
+}
