+++ src/test/java/com/kickball/stadium/service/StadiumServiceMockTest.java
@@ -1,0 +1,38 @@
+package com.kickball.stadium.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.BDDMockito.*;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import com.kickball.stadium.domain.Stadium;
+import com.kickball.stadium.dto.request.StadiumSaveRequest;
+import com.kickball.stadium.dto.response.StadiumResponse;
+import com.kickball.stadium.fixture.StadiumFixture;
+import com.kickball.stadium.repository.StadiumRepository;
+
+@ExtendWith(MockitoExtension.class)
+class StadiumServiceMockTest {
+
+	@Mock
+	private StadiumRepository stadiumRepository;
+
+	@InjectMocks
+	private StadiumService stadiumService;
+
+	@Test
+	void 구장_생성_테스트() {
+		//given
+		StadiumSaveRequest request = StadiumFixture.getStadiumSaveRequest();
+		Stadium stadium = request.toEntity();
+		given(stadiumRepository.save(any(Stadium.class))).willReturn(stadium);
+
+		//when & then
+		StadiumResponse response = assertDoesNotThrow(() -> stadiumService.saveStadium(request));
+	}
+
+}
